Add compatibility for EditorTime
This allows KAC alarms to fire properly when EditorTime is present and
causing time to pass while in the VAB/SPH.

There is one known issue with this change: if the alarm is set to pause
the game, time stops correctly, but the VAB/SPH camera breaks and no
longer moves or rotates. Exiting/re-entering the building is sufficient
to get things working again.