Fix erroneous blocking of input.
The problem might be limited to linux, but Event.current.mousePosition gets
inverted for key-press events causing KAC to set input locks when a key is
pressed with the mouse in a region of the screen that would include a KAC
window when the Y axis is inverted. While Input.mousePosition is also
inverted, it is stable.