Use correct functions for ascending/descending node calculations
It seems KSP 1.12 added functions to Orbit with the same name as KAC
extensions, thus the extensions were no longer called. This resulted in
great confusion about where the AN vector pointed, its true anomaly, and
the time of that true anomaly. This makes use of the functions that are
provided by the Orbit class.